Rename `YamsError` to `YamlError`
